requirements.txt
langchain
langgraph
langchain-core
langchain-community
pandas
numpy
scikit-learn
python-dotenv
langchain-openai
langchain-groq
Ipython
requests
beautifulsoup4
faiss-cpu
nomic
sentence-transformers
tf-keras
hf_xet
autoscraper
lxml
selenium
tabulate
streamlit